+++ cfspopcon/formulas/energy_confinement/__init__.py
@@ -4,7 +4,6 @@ from .energy_confinement_time import (
     calc_energy_confinement_time_from_scaling,
     calc_energy_confinement_time_from_stored_energy_and_input_power,
     calc_H98y2,
-    calc_power_balance_from_input_P_aux,
 )
 from .plasma_stored_energy import calc_plasma_stored_energy
 from .read_energy_confinement_scalings import ConfinementScaling, read_confinement_scalings
@@ -20,7 +19,6 @@ __all__ = [
     "calc_energy_confinement_time_from_scaling",
     "calc_energy_confinement_time_from_stored_energy_and_input_power",
     "calc_plasma_stored_energy",
-    "calc_power_balance_from_input_P_aux",
     "read_confinement_scalings",
     "solve_energy_confinement_scaling_for_input_power",
     "switch_to_L_mode_confinement_below_threshold",